+++ changelog.md
@@ -5,6 +5,8 @@ changes will be documented in this file.
 
 ## [unreleased]
 
+- Rename --repeat flag to --count
+
 ## [0.11.0] 2024-08-06
 
 - Disconnect client if Server.maxQoS is exceeded